@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { TrainingService } from './training.service';
 import * as fromTraining from './training.reducer';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 
 @Component({
